refactor(schemas): hoist shared author helpers; add Prim::author_relationship_targets
`author_float` / `author_bool` / `author_uniform_token` /
`author_uniform_token_vec` / `varying_attribute` / `author_rel_targets`
were duplicated across `schemas/{geom,lux,physics,skel}/author/common.rs`
with identical bodies. Hoist them into one `schemas/common` module so a
behavior tweak (e.g. variability default) lands in one place.

`author_rel_targets` is a generic relationship-targets shortcut — promote
its body to `Prim::author_relationship_targets` so any composed-prim
caller, not just schema authoring, can reach it. The `schemas/common`
version becomes a thin path-based wrapper for the existing
`(stage, prim_path)` callers; it will go away once Authors carry a
`Prim` handle (next commit).

`Prim::new` widens from `pub(super)` to `pub(crate)` to let
`schemas/common` build Prim handles for those wrappers.

Net -29 lines, every schema family's `author/common.rs` shrinks to its
genuinely family-specific helpers.
